# docs/openapi.yaml
openapi: 3.0.0
info:
  title: NEUCourse API
  version: 1.0.0
  description: >
    REST API for NEUCourse — Northeastern University degree planning and scheduling.
    All authenticated endpoints require a Firebase ID token in the
    `Authorization: Bearer <token>` header. Course endpoints are public.
servers:
  - url: /api/v1
    description: Production (Vercel)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token obtained via getIdToken()
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    Course:
      type: object
      properties:
        id:
          type: string
          example: CS3500
        subject:
          type: string
          example: CS
        number:
          type: string
          example: "3500"
        name:
          type: string
          example: Object-Oriented Design
        description:
          type: string
        prereqs:
          type: array
          items:
            type: string
        coreqs:
          type: array
          items:
            type: string
        sections:
          type: array
          items:
            type: object
            properties:
              crn:
                type: string
              meetingTimes:
                type: string
                example: MWF 10:30am – 11:35am
              rooms:
                type: string
              professor:
                type: string
              seats:
                type: number
              campus:
                type: string
...
paths:
...
  /courses:
    get:
      tags: [Courses]
      summary: Search / list courses
      description: >
        Returns up to 50 pre-scraped NEU courses. All query params are optional.
        Results are filtered from an in-memory cache (refreshed hourly).
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Free-text search across courseId, name, and subject.
        - in: query
          name: subject
          schema:
            type: string
          description: Exact subject code, e.g. CS or MATH.
        - in: query
          name: minLevel
          schema:
            type: integer
          description: Minimum course number, e.g. 3000.
        - in: query
          name: maxLevel
          schema:
            type: integer
          description: Maximum course number, e.g. 4999.
      responses:
        "200":
          description: Array of matching courses (max 50)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Course"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /courses/{courseId}:
    get:
      tags: [Courses]
      summary: Get a single course
      description: Returns the full Firestore document for a specific course by its ID.
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
          example: CS3500
      responses:
        "200":
          description: Course data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Course"
        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Plans (authenticated) ────────────────────────────────────────────
...